Moved HTTP server to pkg/server, and added Gin-Gonic
