+++ cmd/config_schema.go
@@ -18,7 +18,6 @@
 package cmd
 
 import (
-	"RiskIdent/jelease/pkg/util"
 	"encoding/json"
 	"fmt"
 	"reflect"
